Extract todo summary projection helper
